Install nimble, allow caching

+++ travis-setup.sh
@@ -1,17 +1,33 @@
 #!/bin/bash
 
-git clone -b devel --depth 1 git://github.com/nim-lang/nim nim
-cd nim
+set -e
+set -o pipefail
 
-git clone --depth 1 git://github.com/nim-lang/csources csources/
-cd csources
+# If Nim and nimble are still cached from the last time
+if [ -x nim/bin/nim ]; then
 
-sh build.sh
-cd ..
+    cd nim
+    git fetch origin
 
-rm -rf csources
+    if ! git merge FETCH_HEAD | grep "Already up-to-date"; then
+        ./bin/nim c koch
+        ./koch boot -d:release
+        ./kock nimble
+    fi
 
-bin/nim c koch
-./koch boot -d:release
+# Download nim from scratch and compile it
+else
+
+    git clone -b devel --depth 1 git://github.com/nim-lang/nim nim
+    cd nim
+
+    git clone --depth 1 git://github.com/nim-lang/csources csources/
+    (cd csources && sh build.sh)
+
+    rm -rf csources
+    bin/nim c koch
+    ./koch boot -d:release
+    ./kock nimble
+fi
 
 cd ..
